+++ db4e/App.py
@@ -14,7 +14,7 @@ from dataclasses import dataclass, field, fields
 from importlib import metadata
 from textual.app import App
 from textual.theme import Theme as TextualTheme
-from textual.widgets import Footer
+from textual.widgets import RadioSet, RadioButton
 from textual.containers import Vertical
 from rich.theme import Theme as RichTheme
 from rich.traceback import Traceback
@@ -29,29 +29,30 @@ except Exception:
 
 from db4e.Widgets.TopBar import TopBar
 from db4e.Widgets.Clock import Clock
-from db4e.Widgets.DetailPane import DetailPane
 from db4e.Widgets.NavPane import NavPane
 from db4e.Modules.ConfigMgr import ConfigMgr, Config
 from db4e.Modules.DeploymentMgr import DeploymentMgr
 from db4e.Modules.PaneCatalogue import PaneCatalogue
 from db4e.Modules.PaneMgr import PaneMgr
 from db4e.Modules.InstallMgr import InstallMgr
+from db4e.Modules.Helper import get_radio_map
 from db4e.Modules.MessageRouter import MessageRouter
 from db4e.Messages.SubmitFormData import SubmitFormData
 from db4e.Messages.UpdateTopBar import UpdateTopBar
 from db4e.Messages.RefreshNavPane import RefreshNavPane
 from db4e.Messages.NavLeafSelected import NavLeafSelected
 from db4e.Constants.Fields import (
-    COLORTERM_ENVIRON_FIELD, COMPONENT_FIELD, DB4E_FIELD, MONEROD_FIELD, P2POOL_FIELD,
-    REMOTE_FIELD, TERM_ENVIRON_FIELD, TO_MODULE_FIELD, 
-    TO_METHOD_FIELD
+    COLORTERM_ENVIRON_FIELD, COMPONENT_FIELD, DB4E_FIELD, INSTANCE_FIELD, MONEROD_FIELD, 
+    P2POOL_FIELD, P2POOL_ID_FIELD, P2POOL_INSTANCE, RADIO_MAP, REMOTE_FIELD, TERM_ENVIRON_FIELD, 
+    TO_MODULE_FIELD, TO_METHOD_FIELD, XMRIG_FIELD
 )
 from db4e.Constants.Labels import (
-    DB4E_LABEL, DEPLOYMENTS_LABEL, MONEROD_SHORT_LABEL, NEW_LABEL, P2POOL_SHORT_LABEL
+    DB4E_LABEL, DEPLOYMENTS_LABEL, MONEROD_SHORT_LABEL, NEW_LABEL, P2POOL_SHORT_LABEL,
+    XMRIG_SHORT_LABEL
 )
 from db4e.Constants.Panes import (
     DB4E_PANE, MONEROD_REMOTE_PANE, NEW_MONEROD_TYPE_PANE, NEW_P2POOL_TYPE_PANE,
-    P2POOL_REMOTE_PANE, 
+    NEW_XMRIG_PANE, P2POOL_REMOTE_PANE, XMRIG_PANE
 )
 from db4e.Constants.Defaults import (
     APP_TITLE_DEFAULT, COLORTERM_DEFAULT, CSS_PATH_DEFAULT, TERM_DEFAULT
@@ -93,7 +94,7 @@ class Db4EApp(App):
     def on_nav_leaf_selected(self, message: NavLeafSelected) -> None:
         category = message.parent
         instance = message.leaf
-        #print(f"App:on_nav_leav_selected(): {category}/{instance}")
+        #print(f"Db4eApp:on_nav_leaf_selected(): {category}/{instance}")
         if category == DEPLOYMENTS_LABEL and instance == DB4E_LABEL:
             db4e_data = self.depl_mgr.get_deployment(DB4E_FIELD)
             self.pane_mgr.set_pane(name=DB4E_PANE, data=db4e_data)
@@ -109,11 +110,26 @@ class Db4EApp(App):
 
         elif category == P2POOL_SHORT_LABEL and instance == NEW_LABEL:
             self.pane_mgr.set_pane(name=NEW_P2POOL_TYPE_PANE)
+
         elif category == P2POOL_SHORT_LABEL:
             p2pool_data = self.depl_mgr.get_deployment_by_instance(
                 component=P2POOL_FIELD, instance=instance)
             if p2pool_data[REMOTE_FIELD]:
                 self.pane_mgr.set_pane(name=P2POOL_REMOTE_PANE, data=p2pool_data)
+
+        elif category == XMRIG_SHORT_LABEL and instance == NEW_LABEL:
+            rec_data = {COMPONENT_FIELD: XMRIG_FIELD, REMOTE_FIELD: False}
+            xmrig_data = self.depl_mgr.get_new_rec(rec_data=rec_data)
+            xmrig_data[P2POOL_FIELD] = self.depl_mgr.get_deployment_ids_and_instances(P2POOL_FIELD)
+            self.pane_mgr.set_pane(name=NEW_XMRIG_PANE, data=xmrig_data)
+
+        elif category == XMRIG_SHORT_LABEL:
+            xmrig_data = self.depl_mgr.get_deployment_by_instance(
+                component=XMRIG_FIELD, instance=instance)
+            xmrig_data[RADIO_MAP] = get_radio_map(rec=xmrig_data, depl_mgr=self.depl_mgr)
+            p2pool_rec = self.depl_mgr.get_deployment_by_id(xmrig_data[P2POOL_ID_FIELD])
+            xmrig_data[P2POOL_INSTANCE] = p2pool_rec[INSTANCE_FIELD]
+            self.pane_mgr.set_pane(name=XMRIG_PANE, data=xmrig_data)
 
     # Exit the app
     def on_quit(self) -> None: